+++ src/lib/components/index.ts
@@ -36,3 +36,6 @@ export {
   SelectTrigger,
   SelectValue,
 }
+
+import { Textarea, type TextareaProps } from './textarea'
+export { Textarea, type TextareaProps }