+++ src/lib/hooks/index.ts
@@ -1,3 +1,4 @@
 import { useScreenSize } from './useScreenSize'
+import { useFetch } from './useFetch'
 
-export { useScreenSize }
+export { useScreenSize, useFetch }