+++ src/misc/Examples.tsx
@@ -2,6 +2,7 @@ import { useState } from 'react'
 import { Button, Input } from '../lib'
 import { ICONS } from '../config'
 import { useScreenSize } from '../lib/hooks/useScreenSize'
+import { useFetch } from '../lib/hooks/useFetch'
 
 export const ButtonAndInputExample = () => {
   const [name, setName] = useState<string>('')
@@ -67,6 +68,44 @@ export const UseScreenSizeExample = () => {
         <li>
           <strong>Height:</strong> <span className="font-mono">{screenSize?.height}</span>
         </li>
+      </ul>
+    </div>
+  )
+}
+
+export const UseFetchhookExample = () => {
+  const { data, error, loading } = useFetch('https://api.ipify.org?format=json')
+
+  return (
+    <div className="flex flex-col gap-6 items-center justify-center mx-auto">
+      <h3 className="text-3xl font-bold">Random API Fetch</h3>
+      <p>
+        Fetching your IP address from{' '}
+        <a
+          href="https://api.ipify.org/"
+          className="text-blue-500 hover:underline hover:text-blue-600 focus:text-blue-700"
+          target="_blank"
+          rel="noreferrer"
+        >
+          api.ipify.org
+        </a>
+      </p>
+      <ul>
+        {loading ? (
+          <li>
+            <span>Loading...</span>
+          </li>
+        ) : (
+          <pre className="font-mono font-light bg-gray-100/50 px-2 py-2 rounded-md">
+            {JSON.stringify(data, null, 2)}
+          </pre>
+        )}
+        {error && (
+          <li className="text-red-600">
+            <strong>Error:</strong>
+            <pre className="font-mono font-light bg-red-100/50 px-2 py-2 rounded-md">{error}</pre>
+          </li>
+        )}
       </ul>
     </div>
   )